+++ include/wex/ex/command-parser.h
@@ -7,12 +7,15 @@
 
 #pragma once
 
-#include <wex/ex/ex.h>
+#include <string>
+#include <wex/ex/command-parser-data.h>
 
 namespace wex
 {
+class ex;
+
 /// This class offers the ex command parser.
-class command_parser
+class command_parser : public command_parser_data
 {
 public:
   /// The possible ex address types.
@@ -44,19 +47,8 @@ public:
     /// or continue parsing address or addressrange (PARSE)
     parse_t type = parse_t::PARSE);
 
-  /// The command.
-  /// mostly a one letter string like "z" for adjust_window
-  /// mostly a one letter string like "p" for print
-  const std::string& command() const { return m_cmd; }
-
   /// Returns true if parsing was ok.
   bool is_ok() const { return m_is_ok; }
-
-  /// The range.
-  const std::string& range() const { return m_range; }
-
-  /// The text (rest), not the original supplied text.
-  const std::string& text() const { return m_text; }
 
   /// The type that is parsed.
   address_t type() const { return m_type; }
@@ -66,9 +58,8 @@ private:
   bool parse_other();
   bool parse_selection();
 
-  bool        m_is_ok{false};
-  std::string m_cmd, m_range, m_text;
-  address_t   m_type{address_t::NO_ADDR};
+  bool      m_is_ok{false};
+  address_t m_type{address_t::NO_ADDR};
 
   ex* m_ex;
 };