+++ src/ex/global-env.h
@@ -26,6 +26,9 @@ public:
   /// the global command.
   explicit global_env(const addressrange& ar);
 
+  /// Destructor.
+  ~global_env();
+
   /// Runs the global commands using specified data.
   /// Returns false if an error occurred.
   bool global(const data::substitute& data);
@@ -37,19 +40,23 @@ public:
   auto hits() const { return m_hits; }
 
 private:
-  bool command(const block_lines& block, const std::string& text) const;
-  bool for_each(const block_lines& match) const;
+  bool command(const block_lines& block, const std::string& text);
+  bool for_each(const block_lines& match);
   bool process(const block_lines& block);
   bool process_inverse(
     const addressrange_mark& am,
     const block_lines&       block,
     block_lines&             inverse);
+  void               skip(const std::string& info);
+  std::optional<int> skip_get_line() const;
 
   const addressrange m_ar;
 
   std::vector<std::string> m_commands;
+  std::set<int>            m_lines_skip;
 
   int          m_hits{0};
+  char         m_marker{0};
   bool         m_recursive{false};
   ex*          m_ex;
   syntax::stc* m_stc;