+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - fixed recording increment macro
 - fixed git diff processing error
 - fixed possible crash when finding previous char using ; or ,
+- fixed global move and copy commands
 
 ## [25.10] - 2025-09-21